refactor: separate express app logic from server listen
